11.11.2017(2)

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -2,6 +2,6 @@ package com.kodilla.testing.shape;
 
 public interface Shape {
     public String  getShapeName ();
-    public int getField ();
+    public double getField ();
 
 }
