+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Square.java
@@ -15,7 +15,7 @@ public class Square implements Shape {
     }
 
     @Override
-    public int getField() {
+    public double getField() {
         return side * side;
     }
 }
